refactor(lexer): extract state functions
